refactor(matrices): move kind, todok and values ro RepMatrix
